added softmax
